+++ src/intervals/factories/BothOpenedIntervalFactory.java
@@ -5,7 +5,7 @@ import intervals.entities.BothOpenedInterval;
 import intervals.entities.Interval;
 
 public class BothOpenedIntervalFactory extends IntervalFactory {
-	public Interval getInterval(double minimum, double maximum, Opening opening) {
+	public Interval getInterval(int minimum, int maximum, Opening opening) {
 		return BothOpenedInterval.create(minimum, maximum, opening);
 	}
 }
