+++ src/intervals/factories/LeftOpenedIntervalFactory.java
@@ -5,7 +5,7 @@ import intervals.entities.Interval;
 import intervals.entities.LeftOpenedInterval;
 
 public class LeftOpenedIntervalFactory extends IntervalFactory {
-	public Interval getInterval(double minimum, double maximum, Opening opening) {
+	public Interval getInterval(int minimum, int maximum, Opening opening) {
 		return LeftOpenedInterval.create(minimum, maximum, opening);
 	}
 }
