+++ src/intervals/factories/RightOpenedIntervalFactory.java
@@ -5,7 +5,7 @@ import intervals.entities.Interval;
 import intervals.entities.RightOpenedInterval;
 
 public class RightOpenedIntervalFactory extends IntervalFactory {
-	public Interval getInterval(double minimum, double maximum, Opening opening) {
+	public Interval getInterval(int minimum, int maximum, Opening opening) {
 		return RightOpenedInterval.create(minimum, maximum, opening);
 	}
 }
